+++ src/test/java/org/folio/processing/mapping/functions/NormalizationFunctionTest.java
@@ -16,6 +16,8 @@ import org.folio.AlternativeTitleType;
 import org.folio.IssuanceMode;
 import org.folio.HoldingsType;
 import org.folio.CallNumberType;
+import org.folio.SubjectSource;
+import org.folio.SubjectType;
 import org.folio.processing.mapping.defaultmapper.processor.RuleExecutionContext;
 import org.folio.processing.mapping.defaultmapper.processor.parameters.MappingParameters;
 import org.junit.Test;
@@ -878,6 +880,125 @@ public class NormalizationFunctionTest {
     var actualInstanceNoteTypeId = runFunction("set_authority_note_type_id", context);
     // then
     assertEquals(STUB_FIELD_TYPE_ID, actualInstanceNoteTypeId);
+  }
+
+
+  @Test
+  public void SET_SUBJECT_SOURCE_ID_shouldReturnExpectedResult() {
+    // given
+    var expectedSubjectSourceId = UUID.randomUUID().toString();
+    var subjectSource = new SubjectSource()
+      .withId(expectedSubjectSourceId)
+      .withName("Medical Subject Headings");
+    var context = new RuleExecutionContext();
+    context.setMappingParameters(new MappingParameters().withSubjectSources(Collections.singletonList(subjectSource)));
+    context.setRuleParameter(new JsonObject().put("name", "Medical Subject Headings"));
+    // when
+    var actualInstanceSubjectSourceId = runFunction("set_subject_source_id", context);
+    // then
+    assertEquals(expectedSubjectSourceId, actualInstanceSubjectSourceId);
+  }
+
+  @Test
+  public void SET_SUBJECT_SOURCE_ID_shouldReturnEmptyIfNoMappingsSpecified() {
+    // given
+    var context = new RuleExecutionContext();
+    context.setMappingParameters(new MappingParameters());
+    context.setRuleParameter(new JsonObject().put("name", "Medical Subject Headings"));
+    // when
+    var actualSubjectSourceId = runFunction("set_subject_source_id", context);
+    // then
+    assertEquals(EMPTY_STRING, actualSubjectSourceId);
+  }
+
+  @Test
+  public void SET_SUBJECT_SOURCE_ID_shouldReturnEmptyIfNoNameSpecified() {
+    // given
+    var expectedSubjectSourceId = UUID.randomUUID().toString();
+    var subjectSource = new SubjectSource()
+      .withId(expectedSubjectSourceId)
+      .withName("Summary");
+    var context = new RuleExecutionContext();
+    context.setMappingParameters(new MappingParameters().withSubjectSources(Collections.singletonList(subjectSource)));
+    context.setRuleParameter(new JsonObject());
+    // when
+    var actualSubjectSourceId = runFunction("set_subject_source_id", context);
+    // then
+    assertEquals(EMPTY_STRING, actualSubjectSourceId);
+  }
+
+  @Test
+  public void SET_SUBJECT_SOURCE_ID_shouldReturnEmptyIfNoMatchingMappingSpecified() {
+    // given
+    var subjectSource = new SubjectSource()
+      .withId(UUID.randomUUID().toString())
+      .withName("Medical Subject Headings");
+    var context = new RuleExecutionContext();
+    context.setMappingParameters(new MappingParameters().withSubjectSources(Collections.singletonList(subjectSource)));
+    context.setRuleParameter(new JsonObject().put("name", "General Subject"));
+    // when
+    var actualSubjectSource = runFunction("set_subject_source_id", context);
+    // then
+    assertEquals(EMPTY_STRING, actualSubjectSource);
+  }
+
+  @Test
+  public void SET_SUBJECT_TYPE_ID_shouldReturnExpectedResult() {
+    // given
+    var expectedSubjectTypeId = UUID.randomUUID().toString();
+    var subjectType = new SubjectType()
+      .withId(expectedSubjectTypeId)
+      .withName("Personal name");
+    var context = new RuleExecutionContext();
+    context.setMappingParameters(new MappingParameters().withSubjectTypes(Collections.singletonList(subjectType)));
+    context.setRuleParameter(new JsonObject().put("name", "Personal name"));
+    // when
+    var actualInstanceSubjectTypeId = runFunction("set_subject_type_id", context);
+    // then
+    assertEquals(expectedSubjectTypeId, actualInstanceSubjectTypeId);
+  }
+
+  @Test
+  public void SET_SUBJECT_TYPE_ID_shouldReturnEmptyIfNoMappingsSpecified() {
+    // given
+    var context = new RuleExecutionContext();
+    context.setMappingParameters(new MappingParameters());
+    context.setRuleParameter(new JsonObject().put("name", "Personal name"));
+    // when
+    var actualSubjectTypeId = runFunction("set_subject_type_id", context);
+    // then
+    assertEquals(EMPTY_STRING, actualSubjectTypeId);
+  }
+
+  @Test
+  public void SET_SUBJECT_TYPE_ID_shouldReturnEmptyIfNoNameSpecified() {
+    // given
+    var expectedSubjectTypeId = UUID.randomUUID().toString();
+    var subjectType = new SubjectType()
+      .withId(expectedSubjectTypeId)
+      .withName("Personal name");
+    var context = new RuleExecutionContext();
+    context.setMappingParameters(new MappingParameters().withSubjectTypes(Collections.singletonList(subjectType)));
+    context.setRuleParameter(new JsonObject());
+    // when
+    var actualSubjectTypeId = runFunction("set_subject_type_id", context);
+    // then
+    assertEquals(EMPTY_STRING, actualSubjectTypeId);
+  }
+
+  @Test
+  public void SET_SUBJECT_TYPE_ID_shouldReturnEmptyIfNoMatchingMappingSpecified() {
+    // given
+    var subjectType = new SubjectType()
+      .withId(UUID.randomUUID().toString())
+      .withName("Personal name");
+    var context = new RuleExecutionContext();
+    context.setMappingParameters(new MappingParameters().withSubjectTypes(Collections.singletonList(subjectType)));
+    context.setRuleParameter(new JsonObject().put("name", "General Type"));
+    // when
+    var actualSubjectType = runFunction("set_subject_type_id", context);
+    // then
+    assertEquals(EMPTY_STRING, actualSubjectType);
   }
 
   private List<HoldingsType> getHoldingsMappingParameter() {